clcc: remove short variant of version information command line flag
The usage help text suggest a shortened `-v` variant of the `--version`
flag, but it is not actually processed, only the long form. Remove the
shortened form from the help text, as e.g. gcc/g++ also only has `--version`,
and it was conflicting with a flag that seems to mean "verbose" too.